requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.4
sqlparse==0.4.2
